chore: apply more light lints (#2575)
<!--
Please provide as much information as possible about what your PR aims
to do.
PRs with no description will most likely be closed until more
information is provided.
If you're planing on changing fundamental behaviour or add big new
features, please open a GitHub Issue first before starting to work on
it.
If it's not something big and you still want to contact us about it,
feel free to do so !
-->

### Problem description
<!-- Describe the bug that you fixed/feature request that you
implemented, or link to an existing issue describing it -->

### Implementation description
<!-- Explain what you did to correct the problem -->

### Screenshots
<!-- If your change is visual, take a screenshot showing it. Ideally,
make before/after sceenshots -->

### Additional things
<!-- Anything else you would like to say -->

+++ plugins/builtin/source/content/pl_builtin_functions.cpp
@@ -27,7 +27,7 @@ namespace hex::plugin::builtin {
 
                 auto selection = ImHexApi::HexEditor::getSelection();
 
-                return u128(u128(selection->getStartAddress()) << 64 | u128(selection->getSize()));
+                return (u128(selection->getStartAddress()) << 64 | u128(selection->getSize()));
             });
 
             /* add_virtual_file(path, pattern) */
